Rename CON.json
I would like to use the Lore Seeker set files in [json-to-mse](https://github.com/fenhl/json-to-mse).
Since MSE is a Windows program, having everything work on Windows is a top priority.
However, since the set file for *Conflux* is named `CON.json`, which is a reserved filename, the repository can't be cloned on Windows, and if it's cloned via WSL, the file can't be opened.
This PR renames the file to `CON_.json`, and adjusts the indexer to ignore the filenames and instead get the set code from the JSON data.